% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tax.R
\name{get_tax}
\alias{get_tax}
\alias{set_tax}
\alias{set_tax.taxlist}
\alias{set_tax,taxlist-method}
\title{Get or set a global taxonomy}
\usage{
set_tax(tax)

\method{set_tax}{taxlist}(tax)

get_tax()
}
\arguments{
\item{tax}{A \linkS4class{taxlist} object set as standard taxonomy.}
}
\description{
For convenience and to simplify the call of name formats, the containing
object (a \linkS4class{taxlist} object) will be handled internally by all
functions.

A standard taxonomic list is already set in this package and can be retrieved
or replaced.
}
\examples{
## Get the current taxonomy
get_tax()
}
