% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn_aun.R
\name{tn_aun}
\alias{tn_aun}
\alias{tn_fun}
\alias{tn_funa}
\alias{tn_auna}
\title{Standard formats for taxonomic names}
\usage{
tn_fun(x, ...)

tn_funa(x, ...)

tn_aun(x, ...)

tn_auna(x, ...)
}
\arguments{
\item{x}{A vector containing selected identifiers for taxon names
(\strong{TaxonUsageID}) included in the internal taxonomic object
(see \code{\link[=set_tax]{set_tax()}}).}

\item{...}{Further arguments passed to \code{\link[=print_name]{print_name()}}.}
}
\description{
These functions work similarly to \code{\link[=tn_an]{tn_an()}} and its alies but the names
are accessed by their \strong{TaxonUsageID} and work also for synonyms.
}
\examples{
## Compare all these outputs
tn_fun(52613)
tn_funa(52613)
tn_aun(52613)
tn_auna(52613)
}
