% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tn_pan.R
\name{tn_pan}
\alias{tn_pan}
\alias{tn_pfn}
\alias{tn_pfna}
\alias{tn_pana}
\alias{tn_pfnas}
\title{Print formatted names for parents}
\usage{
tn_pfn(x, level, ...)

tn_pfna(x, level, ...)

tn_pan(x, level, ...)

tn_pana(x, level, ...)

tn_pfnas(x, level, ...)
}
\arguments{
\item{x}{A vector containing selected identifiers for taxon concepts
(\strong{TaxonConceptID}) included in the internal taxonomic object
(see \code{\link[=set_tax]{set_tax()}}).}

\item{level}{A character value indicating the taxon rank set for parent
taxa. I is passed to \code{\link[=parents]{parents()}}.}

\item{...}{Further arguments passed to \code{\link[=print_name]{print_name()}}.}
}
\description{
Displaying names of parent taxa.
}
\examples{
## Compare all these outputs
tn_pfn(13, "family")
tn_pfna(13, "family")
tn_pfnas(13, "family")
tn_pan(13, "family")
tn_pana(13, "family")
}
